Fixed custom arguments for commandline-specified testrunners
The testrunner that is used can define and `add_arguments()` method to
specify additional commandline arguments for the `test` management
command. To make this work even when the testrunner to be used is
specified on the commandline, the `test` command prescans the
commandline for the `--testrunner` option, to call the runner's
`add_arguments()` before the actual commandline parsing happens.

Previously, the prescanning only considered options looking like
`--testrunner=foo`, while the real parsing also accepts options like
`--testrunner foo`. In the latter case, additional arguments defined by
the testrunner would not be accepted.

This commit fixes this by checking for both argument styles. The
scanning is moved to a separate function for clarity, and to prepare for
possible reuse later.